refactor: move audio info into its own component
